Upgraded bundled Java to 25+36 (Windows x86_64, Mac), 25+37 (Windows aarch64)

+++ bootstrap.sh
@@ -21,48 +21,70 @@ pushd "$JDKDIR"
 
 ZULU_URL='https://cdn.azul.com/zulu/bin'
 
-TEMURIN_URL='https://github.com/adoptium/temurin24-binaries/releases/download'
-TEMURIN_VERSION=jdk-24.0.2+12
-TEMURIN_FILENAME_VERSION=24.0.2_12
+TEMURIN_URL='https://github.com/adoptium/temurin25-binaries/releases/download'
+TEMURIN_VERSION=jdk-25+36
+TEMURIN_FILENAME_VERSION=25_36
 
-BELLSOFT_URL='https://download.bell-sw.com/java/24.0.2%2B12'
-BELLSOFT_VERSION=24.0.2+12
-BELLSOFT_DIR=jdk-24.0.2
+BELLSOFT_URL='https://download.bell-sw.com/java/25%2B37'
+BELLSOFT_VERSION=25+37
+BELLSOFT_DIR=jdk-25
 
 BELLSOFT_WIN32_URL='https://download.bell-sw.com/java/21.0.8%2B12'
 BELLSOFT_WIN32_VERSION=21.0.8+12
 BELLSOFT_WIN32_DIR=jdk-21.0.8
 
+# TODO check for downloads
+
 # Windows x86_32
 filename="bellsoft-jdk$BELLSOFT_WIN32_VERSION-windows-i586.zip"
-curl -O "$BELLSOFT_WIN32_URL/$filename"
-unzip $filename
-mv $BELLSOFT_WIN32_DIR windows-x86_32
+if [ ! -f "$filename" ] ; then
+  curl -O "$BELLSOFT_WIN32_URL/$filename"
+fi
+if [ ! -d windows-x86_32 ] ; then
+  unzip $filename
+  mv $BELLSOFT_WIN32_DIR windows-x86_32
+fi
 
 # Windows x86_64
-filename="OpenJDK24U-jmods_x64_windows_hotspot_$TEMURIN_FILENAME_VERSION.zip"
-curl -L -O "$TEMURIN_URL/$TEMURIN_VERSION/$filename"
-unzip $filename
-mkdir -p windows-x86_64
-mv $TEMURIN_VERSION-jmods windows-x86_64/jmods
+filename="OpenJDK25U-jmods_x64_windows_hotspot_$TEMURIN_FILENAME_VERSION.zip"
+if [ ! -f "$filename" ] ; then
+  curl -L -O "$TEMURIN_URL/$TEMURIN_VERSION/$filename"
+fi
+if [ ! -d windows-x86_64 ] ; then
+  unzip $filename
+  mkdir -p windows-x86_64
+  mv $TEMURIN_VERSION-jmods windows-x86_64/jmods
+fi
 
 # Windows aarch64
 filename="bellsoft-jdk$BELLSOFT_VERSION-windows-aarch64.zip"
-curl -O "$BELLSOFT_URL/$filename"
-unzip $filename
-mv $BELLSOFT_DIR windows-aarch64
+if [ ! -f "$filename" ] ; then
+  curl -O "$BELLSOFT_URL/$filename"
+fi
+if [ ! -d windows-aarch64 ] ; then
+  unzip $filename
+  mv $BELLSOFT_DIR windows-aarch64
+fi
 
 # MacOS x86_64
-filename="OpenJDK24U-jmods_x64_mac_hotspot_$TEMURIN_FILENAME_VERSION.tar.gz"
-curl -L -O "$TEMURIN_URL/$TEMURIN_VERSION/$filename"
-mkdir -p macos-x86_64/Contents/Home/jmods
-tar -C macos-x86_64/Contents/Home/jmods --strip-components=1 -xvf $filename
+filename="OpenJDK25U-jmods_x64_mac_hotspot_$TEMURIN_FILENAME_VERSION.tar.gz"
+if [ ! -f "$filename" ] ; then
+  curl -L -O "$TEMURIN_URL/$TEMURIN_VERSION/$filename"
+fi
+if [ ! -d macos-x86_64 ] ; then
+  mkdir -p macos-x86_64/Contents/Home/jmods
+  tar -C macos-x86_64/Contents/Home/jmods --strip-components=1 -xvf $filename
+fi
 
 # MacOS aarch64
-filename="OpenJDK24U-jmods_aarch64_mac_hotspot_$TEMURIN_FILENAME_VERSION.tar.gz"
-curl -L -O "$TEMURIN_URL/$TEMURIN_VERSION/$filename"
-mkdir -p macos-aarch64/Contents/Home/jmods
-tar -C macos-aarch64/Contents/Home/jmods --strip-components=1 -xvf $filename
+filename="OpenJDK25U-jmods_aarch64_mac_hotspot_$TEMURIN_FILENAME_VERSION.tar.gz"
+if [ ! -f "$filename" ] ; then
+  curl -L -O "$TEMURIN_URL/$TEMURIN_VERSION/$filename"
+fi
+if [ ! -d macos-aarch64 ] ; then
+  mkdir -p macos-aarch64/Contents/Home/jmods
+  tar -C macos-aarch64/Contents/Home/jmods --strip-components=1 -xvf $filename
+fi
 
 popd
